+++ syscore/genutils.py
@@ -71,6 +71,16 @@ def get_unique_list(somelist: list) -> list:
     return list(set(somelist))
 
 
+def get_unique_list_slow(somelist):
+    uniquelist = []
+
+    for somevalue in somelist:
+        if somevalue not in uniquelist:
+            uniquelist.append(somevalue)
+
+    return uniquelist
+
+
 def round_significant_figures(x: float, figures: int = 3) -> float:
     """
     >>> round_significant_figures(0.0234, 2)
@@ -240,6 +250,13 @@ def if_object_matches_return_empty_string(x, object_to_match):
         return x
 
 
+"""
+    
+    NAMED TUPLES AND OBJECTS
+
+"""
+
+
 def transfer_object_attributes(named_tuple_object, original_object):
     kwargs = dict(
         [
@@ -250,6 +267,13 @@ def transfer_object_attributes(named_tuple_object, original_object):
     new_object = named_tuple_object(**kwargs)
 
     return new_object
+
+
+"""
+
+    COMMON FACTORS
+
+"""
 
 
 def highest_common_factor_for_list(list_of_ints: List[int]) -> int: