+++ sysexecution/algos/algo_original_best.py
@@ -242,7 +242,7 @@ def limit_trade_viable(
     if is_market_about_to_close(data, order=order, log=log):
 
         log.msg(
-            "Market about to close or stack handler nearly finished - doing market order"
+            "Market about to close or stack handler nearly close - doing market order"
         )
         return False