+++ sysobjects/production/backtest_storage.py
@@ -4,8 +4,10 @@ import pandas as pd
 
 from syscore.fileutils import full_filename_for_file_in_home_dir
 from syscore.objects import user_exit, missing_data
-from syscore.interactive import print_menu_of_values_and_get_response
-from syscore.interactively_run_functions import fill_args_and_run_func
+from syscore.interactive_menus import print_menu_of_values_and_get_response
+from syscore.interactively_run_functions import (
+    interactively_input_arguments_for_function,
+)
 
 
 class interactiveBacktest(object):
@@ -165,7 +167,7 @@ class interactiveBacktest(object):
 
     def get_result_of_method_for_stage(self, stage_name, method_name):
         func = self.get_method_for_stage(stage_name, method_name)
-        args, kwargs = fill_args_and_run_func(func, method_name)
+        args, kwargs = interactively_input_arguments_for_function(func, method_name)
         try:
             ans = func(*args, **kwargs)
         except Exception as e: